**Feature or improvement description**
This PR adds a feature that determines if there is a minimal rotation of the wind farm layout such that all turbine coordinates will coincide with the non-rotated turbine's coordinates. For rectangular wind farms, this will be a 90 deg rotation. For triangular wind farms, this is a 120 deg rotation, and for rectangular farms, this is 180 deg. Knowing this rotation, the optimization problem can be reduced to a subset of the initial wind directions (i.e., for rectangular farms, from 0 through 90 deg) and then extrapolated to the remaining wind directions.

**Related issue, if one exists**
None.

**Impacted areas of the software**
Yaw optimization.

**Additional supporting information**
This functionality adds very little overhead, so I am defaulting it to `True`. Additionally, I have rearranged some of the yaw optimization code to work with `_subset` variables. This can later also be used to deal with wind farm clustering or other simplifications without having to constantly overwrite the full variables.

**Test results, if applicable**
Here are the results for 3x3 square wind farm:
![image](https://user-images.githubusercontent.com/22119448/149966203-c4f9ba06-dade-4600-81e0-5225fc9fe0bf.png)

Wind farm power:
![image](https://user-images.githubusercontent.com/22119448/149966278-cdfd44c7-3e89-4601-97a8-5b7a6924c079.png)

Optimal yaw angles (first 4 turbines):
![image](https://user-images.githubusercontent.com/22119448/149966387-c3f9f3d4-60f2-4aa8-8b07-db56d10e32a2.png)

![image](https://user-images.githubusercontent.com/22119448/149966426-2328dfd8-79f7-440e-b9d2-e2e05acffe71.png)

![image](https://user-images.githubusercontent.com/22119448/149966483-f14ccb0b-f543-4cf3-8b43-0b955235fb36.png)

![image](https://user-images.githubusercontent.com/22119448/149966523-e2dfa590-636c-464b-b306-29db0182fce0.png)

![image](https://user-images.githubusercontent.com/22119448/149966605-559eba90-c40c-4eb8-b6e2-ecfc6a0a786a.png)

```
Time without symmetry: 5.9 s
Time with symmetry: 3.3 s

         wd_range              turbine_mapping
0     (0.0, 90.0)  [0, 1, 2, 3, 4, 5, 6, 7, 8]
1   (90.0, 180.0)  [6, 3, 0, 7, 4, 1, 8, 5, 2]
2  (180.0, 270.0)  [8, 7, 6, 5, 4, 3, 2, 1, 0]
3  (270.0, 360.0)  [2, 5, 8, 1, 4, 7, 0, 3, 6]
```

And the results with v2.4 are identical too:
![image](https://user-images.githubusercontent.com/22119448/149967100-6a06aced-2ec5-42a1-8b0d-87eba7c0ca12.png)

![image](https://user-images.githubusercontent.com/22119448/149967133-5babf7d3-0f1a-482d-9f97-70a464bbb16f.png)

Additional timings for the square benchmarking layouts, in comparison to the exact same optimization with `exploit_layout_symmetry=False`:
```
[N=25] Time without symmetry: 98.8 s
[N=25] Time with symmetry: 35.6 s (2.8x speedup)

[N=49] Time without symmetry: 661.2 s
[N=49] Time with symmetry: 195.4 s (3.4x speedup)

[N=81] Time without symmetry: 3056 s
[N=81] Time with symmetry: 848.9 s (3.6x speedup)
```

<!-- Release checklist:
- Update the version in
    - [ ] README.rst
    - [ ] docs/index.rst
    - [ ] floris/VERSION
- [ ] Verify readthedocs builds correctly
- [ ] Create a tag in the NREL/FLORIS repository
-->